drf/requirements.txt
pip==22.3.1
django==4.1
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.4